Move validateConnection to own file
