+++ server/drivers/index.js
@@ -1,4 +1,3 @@
-const utils = require('./utils');
 const renderConnection = require('./render-connection');
 const drivers = require('./drivers');
 const DriverConnection = require('./driver-connection');
@@ -53,55 +52,10 @@ function getDrivers() {
   });
 }
 
-/**
- * Validates connection object based on its driver
- * Unnecessary fields will be stripped out
- * @param {object} connection
- */
-function validateConnection(connection) {
-  const coreFields = ['_id', 'name', 'driver', 'createdDate', 'modifiedDate'];
-  if (!connection.name) {
-    throw new Error('connection.name required');
-  }
-  if (!connection.driver) {
-    throw new Error('connection.driver required');
-  }
-  const driver = drivers[connection.driver];
-  if (!driver) {
-    throw new Error(`driver implementation ${connection.driver} not found`);
-  }
-  const validFields = driver.fields.map(field => field.key).concat(coreFields);
-  const cleanedConnection = validFields.reduce(
-    (cleanedConnection, fieldKey) => {
-      if (connection.hasOwnProperty(fieldKey)) {
-        let value = connection[fieldKey];
-        const fieldDefinition = driver.fields.find(
-          field => field.key === fieldKey
-        );
-
-        // field definition may not exist since
-        // this could be a core field like _id, name
-        if (fieldDefinition) {
-          if (fieldDefinition.formType === 'CHECKBOX') {
-            value = utils.ensureBoolean(value);
-          }
-        }
-
-        cleanedConnection[fieldKey] = value;
-      }
-      return cleanedConnection;
-    },
-    {}
-  );
-
-  return cleanedConnection;
-}
-
 module.exports = {
   getDrivers,
   getSchema,
   renderConnection,
   runQuery,
-  testConnection,
-  validateConnection
+  testConnection
 };
