+++ server/lib/config/index.js
@@ -1,6 +1,6 @@
 const _ = require('lodash');
 const appLog = require('../appLog');
-const drivers = require('../../drivers');
+const validateConnection = require('../validate-connection');
 const fromDefault = require('./fromDefault');
 const fromEnv = require('./fromEnv');
 const fromCli = require('./fromCli');
@@ -140,7 +140,7 @@ class Config {
       try {
         let connection = connectionsMap[id];
         connection._id = id;
-        connection = drivers.validateConnection(connection);
+        connection = validateConnection(connection);
         connection.editable = false;
         connectionsFromConfig.push(connection);
       } catch (error) {
